+++ docs/conf.py
@@ -55,9 +55,9 @@ author = 'Ryuichiro Nakato'
 # built documents.
 #
 # The short X.Y version.
-version = '0.10'
+version = '0.13'
 # The full version, including alpha/beta/rc tags.
-release = '0.10.7'
+release = '0.13.0'
 
 numfig = True