+++ tutorial/yeast/Quickstart.sh
@@ -1,5 +1,5 @@
-sing="singularity exec --bind /work,/work2 /work/SingularityImages/churros.0.9.3.sif"
-#sing="singularity exec churros.sif"
+#sing="singularity exec --bind /work,/work2,/work3 /work3/SingularityImages/churros.0.13.0.sif"
+sing="singularity exec churros.sif"
 
 build=sacCer3
 Ddir=Referencedata_$build